+++ src/App.js
@@ -1,7 +1,7 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom"; // padrão config
-import Home from "./pages/Home.js";
-import Register from "./pages/Register.js";
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'; // padrão config
+import Home from './pages/Home/Home.js';
+import Register from './pages/Register/Register.js';
 
 function App() {
